Improve propellant importing
There were a couple of cases that weren't handled very well in the function that processes the propellant in a motor that is being loaded. First, if the loaded motor had a propellant with the same name as one in the library but different properties, it would correctly append a number to the end. However, it wouldn't check if there was already a numbered copy of the propellant with properties that match, and always just increment the number until it didn't overlap any existing propellant. This commit makes it check if the propellant matches any of the numbered copies. It also checks if there is a propellant in the library with the same properties and a different name, and asks if you'd like to merge them.