Improve .app startup time by switching pyinstaller from onefile to onedir
I was under the impression that because MacOS applications are distributed as single .app files, you have to use the pyinstaller `onefile` option when building them. The application was slow to open because it had to unzip its huge dependencies on every startup. It turns out that there is no such requirement, and that if you are willing to accept a larger application, you can get much faster startup times by using `onedir`. The application is still a single .app file, but it opens much faster because the dependencies aren't zipped into a single file internally.