Fix dark mode on Windows versions before 11
Qt's "windows11" theme supports both dark and light mode, but the "windows" and "windowsvista" themes it uses on versions of Windows prior to Windows 11 only support light mode. This commit makes it so oM overrides the theme to "fusion" if running on a pre-11 Windows machine using dark mode, as it looks approximately like native Windows and supports dark widgets.