Improve burnsim file propellant importing
Burnsim files don't include propellant combustion temperature, but it is a required input to calculate C*, which is, in turn, critical for most calculations oM does. As a stopgap, I made it so the importer would set the temperature to a (in retrospect, unrealistic) default value. However, for files that specify both ISP* and molar mass, we actually have all we need to calculate combustion temperature. This commit makes the importer do so when possible.