Fix some metric units truncating values too much
The metric units for nozzle erosion and burn rate coefficient were such that when reasonable values for these fields were displayed in them, the UI's 6 decimal digit limit meant that most if not all of the sig figs were truncated. If you then applied the changes, the value would be lost. This commit adds new metric units for these fields scaled similarly to the imperial units, hides or removes the bad units, and migrates preferences to the new ones.