Fix recent file error when loading file from command line argument
When loading files via a command line arg, the file manager's load method is called before the recent files menu was created, which caused an error to pop up and the file to not be added to the recent files menu, though it was loaded successfully. This commit makes it so it doesn't try to add files to the menu if it doesn't have a reference to it. They are still added to the recent files list so they will be added to the menu when the menu is created during window creation.